Fix loading extensions
